+++ app/components/sd/index.tsx
@@ -1,0 +1,2 @@
+export * from "./sd";
+export * from "./sd-panel";
